# internal/service/api/swagger/docs.yaml
basePath: /api/v1
definitions:
...
  routes.ErrorResponse:
    description: Standardized error response
    properties:
      error:
        description: Error message.
        example: Bad Request
        type: string
      stackTrace:
        description: Stack trace, only populated in debug mode.
        type: string
    type: object
...
  routes.OpenAPISearchResourcesResponseJson:
    properties:
      incompleteTypes:
        items:
          type: string
        type: array
      items:
        description: |-
          This anonymous shape mirrors schemaapi.SearchResourceSummaryJson while
          avoiding swaggo's inability to resolve same-package nested named types.
        items:
          properties:
            labels:
              additionalProperties:
                type: string
              type: object
            matchedLabels:
              items:
                properties:
                  key:
                    example: name
                    type: string
                  value:
                    example: payments-production
                    type: string
                type: object
              type: array
            name:
              example: production-crm
              type: string
            namespace:
              example: root.acme
              type: string
            resourceId:
              example: cxn_test550e8400abcde
              type: string
            resourceType:
              example: connection
              type: string
            updatedAt:
              type: string
          type: object
        type: array
      truncatedTypes:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: AuthProxy is an open-source, embeddable integration platform-as-a-service
    (iPaaS). It manages the connection lifecycle to 3rd party systems, allowing applications
    to call those systems through an authenticating proxy.
  title: AuthProxy API
  version: "1.0"
paths:
...
  /search/resources:
    get:
      description: Search durable resources by first-class name, label value, and
        exact label selector, or seed a bounded recent-resource cache
      parameters:
      - description: 'Search mode: query or seed'
        in: query
        name: mode
        type: string
      - collectionFormat: multi
        description: Resource types; may be repeated
        in: query
        items:
          type: string
        name: resourceType
        type: array
      - description: Case-insensitive literal resource-name or label-value substring
          (minimum 3 characters)
        in: query
        name: q
        type: string
      - description: Exact Kubernetes-style label selector
        in: query
        name: labelSelector
        type: string
      - description: Namespace matcher
        in: query
        name: namespace
        type: string
      - description: Maximum results (query) or results per type (seed), up to 50
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.OpenAPISearchResourcesResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search Admin resources
      tags:
      - search
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: Bearer <token>)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"